+++ CHANGELOG.rst
@@ -11,6 +11,9 @@ Change Log
 * Minor fix to misc_utils.to_integer to handle float strings.
 * Minor fix to structured_data to accumulate unique resolved_refs across schemas.
 * Changes to structured_data to respect uniqueItems for arrays.
+* Handle no schemas better in structured_data.
+* Added portal_utils.Portal.ping().
+* Minor fix in portal_utils.Portal._uri().
 
 
 8.5.0